requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
SQLAlchemy==2.0.29
pgvector==0.2.5
psycopg2-binary==2.9.9
numpy==1.26.4
pillow==10.3.0
tensorflow==2.16.1
python-multipart==0.0.9
httpx==0.27.0
tqdm==4.66.4